Make action work on linux (#6)
Use osslsigncode to sign binaries on linux as well.

This was slightly painful. The version of openssl installed by default requires specifying the "legacy" flag to work with RC2 certificates, the easiest way I found was to pass the certificate through openssl. This depends on ubuntu version as well, openssl 1.x on 20 works, 3.x on 22 requires this workaround.